Move caml bindings out of libsnark, give each pairing its own lib
