+++ src/camlsnark_c/dune
@@ -56,7 +56,7 @@
        -DUSE_PT_COMPRESSION=OFF \\
        -DUSE_ASM=ON \\
        -DOPT_FLAGS='-ggdb3 -O2' ..
-     make -j$(nproc) ff zm snark
+     make -j$(nproc) ff zm snark_caml_common snark_caml_bn128 snark_caml_mnt298 snark_caml_mnt753
 
      mkdir _stubs_build
 
@@ -66,10 +66,25 @@
      mkdir _stubs_build/libzm
      pushd _stubs_build/libzm/; ar -xv ../../depends/libzm.a; popd
 
-     mkdir _stubs_build/libsnark
-     pushd _stubs_build/libsnark/; ar -xv ../../libsnark/libsnark.a; popd
+     mkdir _stubs_build/libsnark_caml_common
+     pushd _stubs_build/libsnark_caml_common/; ar -xv ../../caml/libsnark_caml_common.a; popd
 
-     ar -rc libcamlsnark_c_stubs.a _stubs_build/libff/*.o _stubs_build/libzm/*.o _stubs_build/libsnark/*.o
+     mkdir _stubs_build/libsnark_caml_bn128
+     pushd _stubs_build/libsnark_caml_bn128/; ar -xv ../../caml/libsnark_caml_bn128.a; popd
+
+     mkdir _stubs_build/libsnark_caml_mnt298
+     pushd _stubs_build/libsnark_caml_mnt298/; ar -xv ../../caml/libsnark_caml_mnt298.a; popd
+
+     mkdir _stubs_build/libsnark_caml_mnt753
+     pushd _stubs_build/libsnark_caml_mnt753/; ar -xv ../../caml/libsnark_caml_mnt753.a; popd
+
+     ar -rc libcamlsnark_c_stubs.a \\
+       _stubs_build/libff/*.o \\
+       _stubs_build/libzm/*.o \\
+       _stubs_build/libsnark_caml_common/*.o \\
+       _stubs_build/libsnark_caml_bn128/*.o \\
+       _stubs_build/libsnark_caml_mnt298/*.o \\
+       _stubs_build/libsnark_caml_mnt753/*.o
    popd
    popd
    mv libsnark-caml/build/libcamlsnark_c_stubs.a .
@@ -95,7 +110,7 @@
          -DCMAKE_LIBRARY_PATH=/usr/local/opt/gmp/lib \\
          -DCMAKE_RULE_MESSAGES:BOOL=OFF \\
          -DCMAKE_VERBOSE_MAKEFILE:BOOL=ON ..
-       make -j$(sysctl -n hw.ncpu) ff zm snark
+       make -j$(sysctl -n hw.ncpu) ff zm snark_caml_common snark_caml_bn128 snark_caml_mnt298 snark_caml_mnt753
 
        mkdir _stubs_build
 
@@ -108,7 +123,26 @@
        mkdir _stubs_build/libsnark
        pushd _stubs_build/libsnark/; ar -xv ../../libsnark/libsnark.a; popd
 
-       libtool -static -o libcamlsnark_c_stubs.a _stubs_build/libff/*.o _stubs_build/libzm/*.o _stubs_build/libsnark/*.o
+       mkdir _stubs_build/libsnark_caml_common
+       pushd _stubs_build/libsnark_caml_common/; ar -xv ../../caml/libsnark_caml_common.a; popd
+
+       mkdir _stubs_build/libsnark_caml_bn128
+       pushd _stubs_build/libsnark_caml_bn128/; ar -xv ../../caml/libsnark_caml_bn128.a; popd
+
+       mkdir _stubs_build/libsnark_caml_mnt298
+       pushd _stubs_build/libsnark_caml_mnt298/; ar -xv ../../caml/libsnark_caml_mnt298.a; popd
+
+       mkdir _stubs_build/libsnark_caml_mnt753
+       pushd _stubs_build/libsnark_caml_mnt753/; ar -xv ../../caml/libsnark_caml_mnt753.a; popd
+
+     libtool -static -o libcamlsnark_c_stubs.a \\
+       _stubs_build/libff/*.o \\
+       _stubs_build/libzm/*.o \\
+       _stubs_build/libsnark_caml_common/*.o \\
+       _stubs_build/libsnark_caml_bn128/*.o \\
+       _stubs_build/libsnark_caml_mnt298/*.o \\
+       _stubs_build/libsnark_caml_mnt753/*.o
+
      popd
    popd
    mv libsnark-caml/build/libcamlsnark_c_stubs.a .
