+++ src/camlsnark_c/libsnark-caml/CMakeLists.txt
@@ -254,6 +254,7 @@ add_custom_target(check COMMAND ${CMAKE_CTEST_COMMAND})
 add_subdirectory(depends)
 add_subdirectory(libsnark)
 add_subdirectory(prover_reference_functions)
+add_subdirectory(caml)
 
 find_package(CUDA)
 if(CUDA_FOUND)
