+++ src/camlsnark_c/libsnark-caml/caml/caml_bn128.cpp
@@ -1,0 +1,3 @@
+#include <caml/caml_bn128.hpp>
+
+#include <caml/caml_curve.cpp.template>
