+++ src/camlsnark_c/libsnark-caml/caml/caml_curve.cpp.template
