+++ src/camlsnark_c/libsnark-caml/caml/caml_mnt4.cpp
@@ -1,0 +1,3 @@
+#include <caml/caml_mnt4.hpp>
+
+#include <caml/caml_curve.cpp.template>
