+++ src/camlsnark_c/libsnark-caml/caml/caml_mnt4753_specific.cpp
@@ -1,5 +1,5 @@
-#include <libsnark/caml/caml_mnt4753.hpp>
-#include <libsnark/caml/caml_mnt.cpp.template>
+#include <caml/caml_mnt4753.hpp>
+#include <caml/caml_mnt.cpp.template>
 
 extern "C" {
 using namespace libsnark;
