+++ src/camlsnark_c/libsnark-caml/caml/common.cpp
