--- src/camlsnark_c/libsnark-caml/libsnark/caml/caml_mnt6.cpp
@@ -1,3 +1,0 @@
-#include <libsnark/caml/caml_mnt6.hpp>
-
-#include <libsnark/caml/caml_curve.cpp.template>
