--- src/camlsnark_c/libsnark-caml/libsnark/caml/caml_mnt6753.cpp
@@ -1,3 +1,0 @@
-#include <libsnark/caml/caml_mnt6753.hpp>
-
-#include <libsnark/caml/caml_curve.cpp.template>
