+++ src/dune
@@ -26,7 +26,7 @@
 (rule
  (targets caml_curve.h.template common.h)
  (deps
-   camlsnark_c/libsnark-caml/libsnark/caml/caml_curve.h.template
-   camlsnark_c/libsnark-caml/libsnark/caml/common.h)
+   camlsnark_c/libsnark-caml/caml/caml_curve.h.template
+   camlsnark_c/libsnark-caml/caml/common.h)
  (action
   (run cp %{deps} .)))
